backend/requirements.txt
fastapi
uvicorn[standard]
openai>=1.3.5
httpx>=0.25.0
python-dotenv
